test: Run with relevant React stable types
This has no effect right now.
Once React 19 is out, we're running tests in both 18 and 19.
The types should match that dimension.

Ideally, we'd be able to test RC types but npm claims `npm install @types/react@npm:types-react@rc` is an invalid comparator.
Putting that in the package.json works but I'm too lazy to write a dedicated script right now.